moved forgotten methods into correct repo
